Profile page refactored (you can now view own profile by default, and others' profiles with user ID in link); friends list redirects to those profiles when clicking on a friend/incoming request
